feat: M2-1121 Move Applet global state from Dashboard to Shared (for the builder purposes)
