+++ src/shared/features/AppletSettings/DataRetentionSetting/DataRetention.types.ts
@@ -1,4 +1,4 @@
-import { RetentionPeriods } from 'shared/state';
+import { RetentionPeriods } from 'shared/types';
 
 export type DataRetentionFormValues = {
   period: RetentionPeriods;
