+++ src/shared/types/index.ts
@@ -2,3 +2,4 @@ export * from './activityItems';
 export * from './table';
 export * from './event';
 export * from './fileSystem';
+export * from './retentionPeriods';
