+++ src/shared/types/retentionPeriods.ts
@@ -1,0 +1,7 @@
+export const enum RetentionPeriods {
+  Indefinitely = 'indefinitely',
+  Days = 'days',
+  Weeks = 'weeks',
+  Months = 'months',
+  Years = 'years',
+}
